Remove CYW43_LWIP=1 from the platform-build.py
The build flag CYW43_LWIP is set to zero when building the
arduinopico framework. But when building the SCSI DaynaPORT emulator
firmware the Python build script was setting it to 1. This meant the
cyw43_state struct had a different definition between the
arduinopico static library and firmware build. Causing
the cy43_state.mac memory address to be different between the
library and the firmware.

The solution taken was to comment out the offending lines in the
platform-build.py script.