ZuluSCSI DaynaPORT now shows connection message
Previously cyw43_cb_tcpip_set_link_up and it's similar functions
were not being called in the ZuluSCSI firmware's
`ZuluSCSI_platform_network.cpp`. To fix this their counter parts in
the `lib/core_wrap.txt` had to be removed and the lwip libraries
had to be added back to `tools/libpico/CMakeLists.txt`. Because even
though lwip was disabled for the cyw43 library, the low level driver
for cyw43 had a dependency on a `pbuf_copy_partial_pbuf`
in `pico-sdk/lib/lwip/src/core/pbuf.c`.